SuggestionDesignApp
Using plotly with an amazing design and animation to clarify the visualization by clicking on the button and  selecting the data we want to focus on or analyse.
Some changes of the slider bar to  make it more attractive and well designed, with the blue color rather than the black and white.